[~] fixed major and minor coding style errors
